backend/functions/config/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
google-cloud-aiplatform==1.*
firebase-admin==6.*
python-dotenv==1.*
google-cloud-texttospeech==2.*
google-cloud-storage==2.*
google-adk
